Started to implement KNN classifier for tiny images

+++ src/main/java/uk/soton/ecs/COMP3204/cw3/groupA/TinyImageFeature.java
@@ -28,9 +28,9 @@ import org.openimaj.feature.FeatureExtractor;
 
 public class TinyImageFeature implements FeatureExtractor<DoubleFV, FImage>{
 	
-	int scaleSize;
+	float scaleSize;
 	
-	public TinyImageFeature(int scaleSize) {
+	public TinyImageFeature(float scaleSize) {
 		this.scaleSize = scaleSize;
 	}
